Build second bundle for log window
